+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Surface transport errors and align docs with behaviour
 - Align create/update body to the live v2 spec
+- Merge --env over template env on create
 
 ## [0.3.0] - 2026-07-30